Move bespoke layout ID to m_extra
Summary:
m_extra is completely unused in ArrayData. For BespokeArray, we can use it for the layout ID.

Now that BespokeArray isn't abusing m_size, we have the option of taking a blood oath stating that we'll always set this field correctly. (As is typical with such contracts, we enforce them in checkInvariants.) In exchange for our souls, we get to make count IR ops Layout Agnostic (TM). We can drop quite a few bytecodes from irgen-bespoke.

Differential Revision: D23468259

fbshipit-source-id: 4921c059e64cced81c4ee33551126df8770f47e8