De-dupe serialization logic
Summary:
I audited our serialization logic for handling the legacy mark, and found that we have an alternate implementation of the "base case" for serialization in ext_std_variable.cpp that does the same thing as VariableSerializer, but only one level deep. If we have to serialize a recursive object, it calls into VariableSerializer.

Why would we duplicate this logic between serialize and VariableSerializer? I assume it's a performance optimization, but if so, we can likely get the same benefit by moving the logic here into VariableSerializer instead. Unsurprisingly, there are a few discrepancies between the top-level serialization logic and the general logic; the top-level logic throws on large strings and handles resources specially. Oh well...serves us right for duplicating code.

Reviewed By: ricklavoie

Differential Revision: D23420488

fbshipit-source-id: 39e4fe9a3cd7de289419e441e0e4f088b64d9eec