Add content of multifile files to error messages when running in `hh_single_type_check`
Summary:
The contents of multifile files of the form:

```
////long/directory/name/class_def.php
<?hh // strict
class Foo {
  public function f(int $x): void {}
}

////another/directory/treq_ext_foo.php
<?hh // strict
trait TReqExtFoo {
  require extends Foo;
}
```

don't get displayed in error messages because the path passed to `Errors.read_lines` is the combination of real file name and the subfile name, e.g. `sample/file/name.php--another/file/name.php`.

This tries to read the file as normal, and if it fails, tries to read it as if it was a multifile.

Before:

{F319350487}

After:

{F319350480}

Reviewed By: Wilfred

Differential Revision: D23471740

fbshipit-source-id: 54da4ee84c84724c9a92008481a2035c4dbb3691