delay registration of Func in reverse data map
Summary: The reverse data map assumes persistent object that won't go away (and takes advantage of that to avoid excessive locking). However, there are short-lived class methods. This diffs tries to delay the registration of methods to a point where the Class is registered, to avoid registering and unregistering frequently.

Reviewed By: ottoni

Differential Revision: D17381014

fbshipit-source-id: c4ff1a77761a3232bd958046bf90e9e3bf065451