support external arguments in governed functions
Summary:
External arguments flow into governed function, but we do not know for sure whether they have the implicit policy.

In practice this means that they are read-only, however the external constraint has some other implications. For example, they cannot be returned or stored, because that would make them mutable.

Reviewed By: mpu

Differential Revision: D23474869

fbshipit-source-id: ce8481965a154091664d99bad18ae8c7942de232