Enum Supertyping [ 1. Parsing ]
Summary:
This is the first diff of a stack that will implement the Enum Supertyping HIP D23060077 (in experimental status).

This diffs updates the AST and parsing to support for the `includes E1, .., En` syntax in enum declarations (and dependencies).

The extension is gated via `enum_supertyping` unstable feature flag.

Typechecking and runtime will be covered in other diffs.  Given the number of support files touched by this simple parser modification, it would be great if this first diff could be reviewed on its own.

Reviewed By: vsiles

Differential Revision: D23447721

fbshipit-source-id: 8f81f5cc68d95d4bfe27fa33d23885d6e33940d6