Restrict arrprov::Tag::line() to Kind::Known
Summary:
I would like to squeeze an arrprov::Tag into 4 bytes in lowptr builds. Doing so will allow us to store tags inside ArrayData, saving memory and improving performance.

Right now, a tag is a pair of a 4-byte lowptr "file" and a 4-byte line. However, most kinds of tags don't use the line. The only ones that do are Tag::Known (provenance in Hack), Tag::RuntimeLocation, and Tag::RuntimeLocationPoison. The two runtime locations are computed statically, so we can just fold the line into the name.

We rename "filename" to "name", because for different tag kinds, the name means different things: either a Hack filename, a C++ filename (or, now, a C++ file-line pair), or an enum classname. We also pull out a few other functions so that code can treat them homogeneously; all of these functions will change in implementation as we compress tags.

WARNING: This diff changes the serialization format for runtime location tags, dropping the int line. I think that's safe because tags are only serialized in the Internal format or the new provenance-preserving Serialize format, which all guarantee that the same binary is used to serialize and serialize; dneiter, please let me know if you're concerned.

Reviewed By: dneiter

Differential Revision: D23585630

fbshipit-source-id: eb310f6e9235a437d67dffaf16ce7dfb79df1dbc