Build `HH\Facts\type_name()`
Summary:
This is a nice-to-have which allows us to remove a bunch of `HH_FIXME`s.

As far as I know, the best way to resolve a string to a `classname<nonnull>` is to call `ArgAssert::isAnyClassname()`. But you'll notice that this uses the unit loader to enforce the class's existence via calling `PHP\is_a()`) and/or `PHP\trait_exists()`, and it has an `HH_FIXME` anyway to coerce the string to a `classname<nonnull>`.

https://www.internalfb.com/intern/diffusion/WWW/browse/master/flib/core/utils/ArgAssert.php?commit=1002639401&lines=1169-1186

Native Facts can avoid the unit loader and return a `classname<nonnull>` without using an `HH_FIXME`.

Reviewed By: dneiter

Differential Revision: D23609953

fbshipit-source-id: a84882f439bef5ba25ab231b23dca37e0caf8aad