Add a "--no-skipif" flag to the test runner.
Summary:
This flag causes "*.skipif" files to be ignored, as if they did not exist.
If the skipif file would have emitted output, then the associated test
will be run instead of being skipped, and otherwise, the test will be
run as usual, but slightly faster.  So this flag is useful to (1) run tests
which are known to pass slightly faster than normal, or (2) to see if
a test might still pass even if any associated skipif was deleted.

Reviewed By: alexeyt

Differential Revision: D23577893

fbshipit-source-id: bb47a8276932275a93a20a0a36a49a59cc38b1f7