Move back unit-cache freeing to original spot
Summary:
When I implemented the Unit prefetcher, I put the logic to
join the prefetcher threads at the same point we freed the unit-caches
during shutdown. However, joining the prefetcher threads needed to be
done before module shutdown (because of some assertions when we
hphp_thread_exit), so I moved up the unit-cache freeing. This turns
out to be incorrect, as we might still have execution happening
earlier. Therefore we'll free Units we're still executing. Move the
unit-cache freeing back to its original point. Join the prefetcher
threads in a separate function, which we'll call earlier.

Reviewed By: mofarrell

Differential Revision: D23622093

fbshipit-source-id: 1d8062fee09d9b608d0c8d81f3e595021ef0f61a