Add raw color output for highlighted error formatter tests
Summary: Adds an environment variable, checked in `Tty.should_color`, to force color/escape codes to be output for error messages even when not directed towards the terminal. This allows the highlighted error format raw output to be checked (whereas previously the escape codes for color/bold/etc. were being eaten up when sent to the `.exp` files).

Reviewed By: Wilfred

Differential Revision: D23489504

fbshipit-source-id: 952e20a9101bdf0ae889f016db7f13d6400d2364