Track logging array monotypedness
Summary:
For logging arrays, we now track the monotypedness of the array over its
lifetime. Upon construction, each logging array is assigned a MonotypeState
that indicates the monotypedness of its keys and values. Value monotypedness is
represented using a lattice of {Empty, SingleType, SingleTypeNullable, Any},
and key monotypedness is represented using a lattice of {Empty, Ints,
StaticStrings, Strings, Any}. For each operation, we log the monotypedness
state of the array and if it was changed by the operation. We then report these
statistics in the final dump.

Differential Revision: D23453488

fbshipit-source-id: 4b9df5f2e623077575459dc5c513f4900b468624