Allow nested markdown delimiters in error messages
Summary:
This adds limited support for nested Markdown delimiters in error messages, e.g.
```
`camel~~C~~ase`
```
renders is underlined with 'C' highlighted in color. Here's an actual example:

{F333221853}

This diff's approach is to accumulate a list of delimiters for each character in a message, recursively trying parse the contents of valid sections. This is not an attempt to cover all the possible edge cases of a full-fledged markdown parser (e.g. there are a lot of rules just related to nesting in the Commonmark [spec](https://spec.commonmark.org/0.29/#emphasis-and-strong-emphasis)), but basic things like the above example should work.

Reviewed By: Wilfred

Differential Revision: D23621091

fbshipit-source-id: c0578c577cc780a8ccf9c5234a567119e14c8ac9