Make error data structure a record with main and supporting messages fields (part 1)
Summary:
Refactors the error type to be a record containing `code` and, instead of just a list of messages, a `claim` a `reasons` field. This is to better convey the implicit assumption that we have in our code, that the list of messages in an error is always non-empty. Now, only the `reasons` list may be non-empty; `claim` must be set in order to build an error.

This raises a few interesting points related to the `try`, `try_with_error`, and `try_with_result` functions, the latter in particular in its dealing with sentinel/fake messages.

I've left `add_list` in for now as its used in a lot of places, but now it calls `make_error` which fails if the implicit condition that the list is non-empty isn't met. This isn't ideal, so the goal is to continue refactoring so `add_list` is unneeded/removed and only `add_error` is used.

Reviewed By: Wilfred

Differential Revision: D23351710

fbshipit-source-id: 8518899762fd048131cb93d0296efa96b22ff1d8