Fix object serialization in arrprov mode
Summary: Objects are serialized by constructing an internal props array, but we don't deserialize these arrays via the generic array code, so we don't handle provenance in them. The right thing to do is to avoid ever serializing "provenance" for this internal array. We can achieve that by converting it (in place) to a dict.

Reviewed By: dneiter

Differential Revision: D23641509

fbshipit-source-id: 9f5090f342192270962fd0be91be8a0e267536f3