Add forceHackArrays serialization flag
Summary:
I've long wanted this option. It will let us round-trip dvarrays through JSON post HADVAs, for example. It's the best way to fix arrprov INI serialization, so it's time to land it.

When this option is used with json_encode, we'll serialize a darray as a JSON map {...}, regardless of the structure of the darray. An empty darray will be serialized as {}, and a list-like darray will be serialized as {"0": ..., "1": ...}

Reviewed By: dneiter

Differential Revision: D23665289

fbshipit-source-id: 6607d4bc18282ee99ce4aa44aecce32c94d2cb6d