Fix ini usage in arrprov mode
Summary:
Outside of a PHP context, we need to guard the creation of any PHP arrays with an ARRPROV_USE_RUNTIME_LOCATION check, and we need to do something about observations. I've opted to just make us use non-observing serialization in these sites: specifically, json_encode with the "force Hack arrays" option.

This change may change debug output slightly, but it's not used for anything else.

Reviewed By: dneiter

Differential Revision: D23665666

fbshipit-source-id: 6ab5f079748b620402b1cf6d9d324752732cc586