introduce <<CanCall>> annotation
Summary:
The CanCall annotation allows governed function to call private functions.

Practically speaking, this is the opposite of <<External>>. It can be used for inputs that are not functions and will make them write-only (external is essentially read-only). However, I'm not sure how useful that is. I anticipate that this will mostly be used for functions, so I think the CanCall name works well. We can change it if others disagree.

Reviewed By: mpu

Differential Revision: D23501775

fbshipit-source-id: 1d2f328fd8f941eac812cc1ae4e10ceec741275a