Make analysis resilient to unknown syntax
Summary:
While we are still working on supporting all of Hack's syntax, it is useful for the analysis to be able to gracefully handle parts of a function that it CAN recognize.

In order to achieve this, I created a new error type called "unknown information flow". Whenever we encounter unknown syntax, we emit this error and then keep going. In prod, we could presumably suppress these errors so that the analysis can successfully run on real source files. This will make the analysis unsound, but it will allow us to move forward.

Reviewed By: mpu

Differential Revision: D23625891

fbshipit-source-id: ea14e9af9966bf1860476aa845524b304d44fde6