JIT provenance tags
Summary:
This diff is the reason why I stopped giving a damn about the cost of hashing to compress provenance tags: the JIT is our ace in the hole. We can burn constant tag IDs into the JIT for all array constructor ops that happen in non-ProvenanceSkipFrame functions.

Instead of passing the tag parameter to the allocation function, we could also do the store in the JIT via a new SetProvenanceTag IR op - but that's more code, and we still need to suppress the tag lookup within the function, so I don't think it's meaningfully better.

Reviewed By: arnabde03

Differential Revision: D23667276

fbshipit-source-id: da6556b154c87c5707d302616ada8b18f549c4a3