Store provenance tags in arrays
Summary:
Rather than hashing the ArrayData* to put the arrprov::Tag in a side table, we store the data on the array itself in the m_extra field. This field is also used by bespoke arrays; the two features are mutually incompatible, so that doesn't pose any problems. (Access to the tag is guarded by a vanilla assertion.)

There's a lot more we could do to simplify things further. For example, we could also store the tag in the AsioExternalThreadEvent struct and eliminate the request-local hash map completely. I'll do that in a follow-up; it has minimal effect on performance. The next step for actual performance wins is to burn tag writes into the JIT. That's also why I read and write it here via an offset - we're going to expose the offset anyway, so I don't care to add any other accessors.

Reviewed By: arnabde03

Differential Revision: D23661416

fbshipit-source-id: 21b5b8ccb80c368589c10f776b7d81fa4e745996