Pack arrprov::Tag into 4 bytes
Summary:
A provenance tag contains both a LowStringPtr and a line number. If we can fit the tag in 4 bytes, we can store the tag inline, within the array, rather than in a separate side table, saving both memory and instructions. However, the easiest way to shrink the tag is to use hashing, which eats into most of the win - it turns a potential 6pp win into a net 2pp win.

I'm going to address the cost of compressing the tag in a follow-up diff, where I add JIT support for setting provenance tags. For now, I'm okay paying the cost.

Reviewed By: arnabde03

Differential Revision: D23644440

fbshipit-source-id: 738c96d52cb682cc24ea1964c0eabf589455d1cf