Treat calls through multiple upper bounds as call through intersection type
Summary:
Currently we reject calls through the upper bounds on a generic parameter if there is more than one signature for the member. For example, we reject
```
interface I {
  public function Bar(mixed $i): int;
}
interface J {
  public function Bar(string $s): arraykey;
}

function ExpectInt(int $a): void {}

class G<Tg as I as J> {
  public function Callit(Tg $x): void {
    ExpectInt($x->Bar('c'));
  }
}
```
We produce an error
```
Typing[4161] You are trying to access the method Bar but there is more than one implementation on a value of generic type Tg [1]
```
This isn't very well worded: there is in fact only one *implementation*, it just happens to satisfy two method signatures.
Also, it's inconsistent with the behaviour on intersection types. For example:
```
function InvokeOnIntersection(mixed $x): void {
  if ($x is I && $x is J) {
    // Here Hack thinks that $x has type (I & J)
    ExpectInt($x->Bar('c'));
  }
}
```
So it turns out that there is no need for the "ambiguous member" error after all. In the implementation, we construct an intersection of upper bounds for the generic parameter, and re-use the code for intersections.

Reviewed By: losvald

Differential Revision: D23597293

fbshipit-source-id: e2cefbe097dd12c2e494f027cc4f2e60e2c247f0