introduce the thread-local static arena
Summary:
This is a bump allocator from a per-thread buffer, which can be used to hold global data that are very rarely freed until end of the process. Each thread has its own buffer to improve locality. For this diff, we put static strings there. Subsequent work can allocate other kinds of data, e.g., hot arrays, using this mechanism as well.
The option Eval.UseTLStaticArena can be used to turn it off, and it will fallback to use lower_malloc/lower_free.

Reviewed By: ottoni

Differential Revision: D23405537

fbshipit-source-id: e77cb4320329765f232e49c31783e70f8f826abf