Log on marked cast to Hack array
Summary:
We need to log when we cast a marked varray to a vec, or a marked darray to a dict. After we flip the big switch, these casts will be no-ops - literally, generating no code at all - so we'll need to insert a separate "array_unmark_legacy" call into our codebase at the points that used to reset the mark.

This diff also adds a new invariant to our array types: a legacy array is always a HAM-safe dvarray (that is, a dvarray pre-HADVAs, and a vec or a dict post-HADVAs). This assertion didn't pass before: we didn't consistently clear the mark on varray->vec casts or darray->dict casts. We now do.

Reviewed By: dneiter

Differential Revision: D23686370

fbshipit-source-id: 0c793d44c22263d7508a29a8ca03c7001be4aec8