Don't mark server as warmed up unless we have emitted some code
Summary:
In jumpstart mode the servers sometimes jump to 100% maturity, and become
"warmed-up" prior to emitting any live tranlsations.  This adds a check to
ensure we have filled at least some portion of the TC, and then leaves it to
the rate checks to ensure we have slowed down emitting code before marking the
server as fully warmed up.

Reviewed By: ottoni

Differential Revision: D23706722

fbshipit-source-id: 092589c2e085f452dce64ba984aac99e7a36187c