type-at-pos-batch only get tast once per file
Summary:
If you did `hh --type-at-pos-batch file.php:1:1 file.php:2:2 file.php:3:3` for a load of positions in the same file, it formerly used to compute the TAST for file.php each time. (that would have been okay if we cached them, but due to a bug we didn't).

This diff changes it to only fetch TASTs once per file.

Reviewed By: hgoldstein

Differential Revision: D23666551

fbshipit-source-id: 3691ee990d83cd766cb4d4f49d0077ecc70f48c9