New-style shape/tuple logging
Summary:
There are multiple issues with shape/tuple behavior. First of all, we restricted tuples to varrays in is/as positions, but not in reified generics positions (and likewise for darrays and shapes). Second, for HAM, our plan of record is to avoid observing arrays passing through shape/tuple type checks, but to log when a Hack array passes through these types, and extract the is_php_array check when that occurs.

In this diff, we introduce logging to solve both of these issues. We also disable provenance logging at these sites. Our logging is tight; we only log if a vec reaches a tuple type-check and would have passed the structural component (and likewise for dicts and shapes).

Reviewed By: dneiter

Differential Revision: D23707655

fbshipit-source-id: 7cb6a60e2409720f5d64d8da6a2708e3937713f1