Fix negative lookup cache bug
Summary: When we look up the decl of a class named Foo and find it, we want to write its filename to our naming-table cache, to avoid IPC round-trips for information we already know. There's a bug here, though--when we look up the decl of a class named Foo and *don't* find it, we record its non-existence in our naming-table cache. This is the correct behavior for functions and global constants, but it's wrong for classes--there may be no *class* named Foo, but there may still be a typedef or record with that name. For type lookups, we should only write to the naming table cache if we get a positive result.

Reviewed By: shiqicao

Differential Revision: D23690702

fbshipit-source-id: 0b86bd221f1b669a0a507db4b8c749f1e7468b3a