Use target profile bytecode offset instead of translation bytecode offset
Summary:
Sometimes the ArrayIterProfile data showed the incorrect offset. This is because I based the code off of `show(SrcKey)`, which was part of the initial proposal before we moved to actual JSON serialization, as it also has the same issue. :)

This patches up the specific use of the print_profiles trace to use the target profile offset.

Differential Revision: D23725947

fbshipit-source-id: b9992f2dc4b0fe57537783575a2a2590e10c57ce