Set tcopt based on __EnableUnstableFeatures
Summary:
File attributes are erased as top level declarations in the NAST, but first they are stapled into the declarations for classes and functions as the fields `{c,f}_file_attributes` .

https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/hphp/hack/src/parser/namespaces.rs?commit=441c91b61239&lines=289-301

Previously, `f_file_attributes` were discarded during Naming; this diff restores them. The attributes are processed again Typing_toplevel, which is a convenient place to toggle typechecker options in the environments before they are handed off to type function bodies.

The upside is that we can test potentially expensive experimental features on a per-file basis in WWW. The actual toggling of options is fairly mechanical, but there is room for improvement in syncing up the Rust `UnstableFeatures` struct with the module in Naming_special_names.

Reviewed By: losvald

Differential Revision: D23725906

fbshipit-source-id: cd88c1247f0ac9ea52f1e1302473a7ebad0ed6e0