New configuration setting to set sample rate for typechecking
Summary:
When testing new typechecker changes on large codebases, the development cycle is very long. This diff lets us reduce the cycle by "sampling" files, only running the type check phase for a full-init check over some specified fraction of them. The sampling is deterministic as it is determined by filename hashes.

We add a new option `typecheck_sample_rate = f` to `.hhconfig`, where `f` is a float that specifies the fraction of files to be typechecked. The default is 1.

Reviewed By: CatherineGasnier

Differential Revision: D23704622

fbshipit-source-id: 0389812e5731d87e413d425028c820e365ba54bd