Allow logging arrays to flow through profiling translations
Summary:
In profiling tracelets, we no longer guard on arrays being vanilla on entry. All arrays that reach a profiling tracelet are either vanilla or logging. For vanilla arrays, we wish to preserve irgen as much as possible to avoid interfering with profiling information. For logging arrays, we wish to interpret all layout-sensitive bytecodes, but within the tracelet they appeared. This will allow us to better track in which tracelet array events occurred.

We do the following:
1) Perform region selection for profiling tracelets as if the tracelet had vanilla constraints.
2) Unspecialize guards for arrays in profiling tracelets so that vanilla guards are not emitted at entry. This occurs in `emitGuards` so that preconditions used for optimized region formation are unaffected.
3) When we translate a layout-sensitive bytecode, branch on if the array is logging or not. In the vanilla branch, translate as usual assuming the array is vanilla. In the other branch, interpret the bytecode.
4) To recover type information after interpreting, emit AssertTypes after the join point for local and stack locations based on the type information in the vanilla branch.

Differential Revision: D23639135

fbshipit-source-id: 6843f2e0d9f400d494a6645c7264f4c08de5f8d3