Record reach event for LoggingArrays which pass a weakened DataTypeSpecialized guard
Summary: We record that a LoggingArray reached a tracelet after it passes its weakened DataTypeSpecialized guards in profiling tracelets. We track each reach event by recording the tracelet ID and the index of the DataTypeSepecialized precondition/weakened guard that the array passed. This data will be used to select which guards to specialize during RTA. Improved output in the dump will be forthcoming.

Differential Revision: D23768101

fbshipit-source-id: 850e58cf6727226abcd8f0db85adc8e47e07f84e