Clean up array-provenance a bit
Summary:
Get rid of the current pointer arithmetic skulduggery to access
m_extra off of an ArrayData--properly friend Tag. This makes things a little
harder to accomplish in `array-provenance.cpp` but we just need to add the
appropriate friendships.

Also kill the `Mode` parameter on `setTag`--essentially everyone specifies
`Mode::Emplace` which is the more permissive one--just make that the behavior.

Differential Revision: D23726787

fbshipit-source-id: 975aeeffc7bf8c417eec5370332b68e75744e704