Kill kHasProvenance bit in ArrayData--m_extra is always initialized
Summary:
We can just use `m_extra == 0` instead of an extra bit in the
arraydata header--this is convenient for cpgo also since if m_extra gets
initialized we can **also** use it as a proxy check for the array being bespoke
or not. Which is great! Yay~!

Also monkey patch the nonsense w/ `tag_slot(...)` in array-provenance.cpp so
the dbgo build doesn't (correctly) determine that the reference it returns is
weird at the time it's written to and optimize the writes away wheeeEEEee~

Differential Revision: D23725521

fbshipit-source-id: e434f45afb726efb347584c47c855a4b0829841f