Fix bug in onFunctionEnter
Summary:
If EnableArgsInBacktraces=false, we may unset or reuse the reified generics local. (We need the same check to include reified generics information in backtraces - see backtrace.cpp.)

At function entry, the frame will always be in the expected state. We can keep the generics around for enters regardless of the flag - we only need to drop them for resume.

Reviewed By: ricklavoie

Differential Revision: D23804685

fbshipit-source-id: 2e889d0ee75677d13556d9cdefd7b3ea92e5e76b