fix per-cont typechecking for finally
Summary:
In the typechecker, we analyze `finally` blocks against all continuations individually in order to get more precise typing information about the locals in each outcome. However, the implementation was broken because it did not use the `Next` continuation when merging all of the envs back together in `typing_try.ml`. This meant that we did not in fact accumulate any information about the locals that we bind in finally blocks!

To fix this, I needed to make a few changes
* When doing a typechecking pass on a finally block, instead of just replacing Next with the cont we are checking, I create a singleton CMap with only the continuation we need. The other continuations are not relevant and in fact introduce imprecise information when merging back together
* When merging the conts back together, instead of taking the `cont` from `cont` I take `Next` from `cont`. We want the Next continuation because that is what has updated information about locals.

The tests I added did not pass before my diff

Reviewed By: CatherineGasnier

Differential Revision: D23758135

fbshipit-source-id: cdac526c96efc0f5fd89a93e65e10faf0edb6853