Add pass to optimize vanilla checks
Summary:
After jjgriego's diffs that initialize m_extra for all arrays, we can play a trick to optimize vanilla type checks: we can treat (m_size, m_extra) as a single quadword, which is always the array's size (for vanilla arrays) and always a negative integer (for bespoke arrays).

For "known offset" MixedArray lookups, we can test the offset against this quadword instead of against m_used and implement both CheckDictOffset and CheckVanilla in one test. Similarly, for PackedArray lookups with constant keys, we can test against this quadword and implement both CheckVecBounds and CheckVanilla in one test.

We call these two new operations CheckDictOffsetLA and CheckVecBoundsLA. The LA suffix stands for "layout-agnostic", like the opcode flags bit. In this diff, we add an optimize pass that rewrites the CFG, inserting these operations to reduce the cost of vanilla checks. This pass does a loose domination analysis, but we don't build a complete dominator tree; for the typical cases we want to rewrite, we only need to analyze 2 or 3 blocks ahead to locate the merge point.

Differential Revision: D23778275

fbshipit-source-id: bcd770aea460cd392ceecfe1fdaee65b6c46ad88