Account for mem to mem copies in spill loop hoisting penalty calculation
Summary:
When calculating whether its profitable to hoist a spill outside of a
loop, account for the fact that we may create memory to memory
copies. If the source of such a copy is live across the copy, the
memory to memory copy is particularly expensive. That's because we
cannot coalesce the copy (because of interference), and memory to
memory copies must be lowered to a load/store pair. So, hoisting the
spill may actually make the copy more expensive. In this situation,
assess a penalty representing the cost of the load/store pair. In
addition, account for instructions which may become nops.

Reviewed By: mofarrell

Differential Revision: D23777873

fbshipit-source-id: 6b6a2aa9ff1dfb9e6808fe2a0ff38038c90b1ead