Make SP equal to FP in DefFrameRelSP
Summary:
If the stack is defined relatively to the frame, there's no advantage to define
SP at an offset from FP, as all SP-relative access uses additional offsets.
There is only a disadvantage that under a register pressure we may have to
spill this derived SP if we were not able to optimize away double offsets.

The stack base still remains the same by setting irSPOff to zero. This will be
revisited by future diffs.

Reviewed By: ricklavoie

Differential Revision: D23808890

fbshipit-source-id: 9db5c2b383b32817268d7c0dae98df74c86b1a1a