Generalize stublogue fixup mechanism
Summary:
Code in stublogue context needs an indirect fixup. This is currently achieved
by manually passing the SyncStublogue option to the call and it is used only by
CheckStackOverflow from the initial section of prologue.

This is inconvenient, as instructions that can throw may operate only in one of
these two modes, or need to be aware which mode they are in and manually set
the correct sync mode.

Let's instead use the marker to track whether we are in the stublogue context
and setup the correct fixup automatically.

Reviewed By: ricklavoie

Differential Revision: D23809454

fbshipit-source-id: 8e13b8580f82cc1ce882b485b5dc21eb939bb3ec