Defer stack initialization to Def*SP instructions
Summary:
The responsibility for stack initialization was spread across FrameState's
constructor and Def*SP instructions. Move it to the Def*SP instructions, which
now contain all information needed to setup all of `sp(env)` pointing to some
location in the stack, `irSPOff` representing the offset of `sp(env)` from the
base of the stack and `bcSPOff` representing the offset of the top of the stack
from the base of the stack.

This diff is a noop, as these offsets were previously configured in the same
way. Makes it possible to properly reconfigure stack in the middle of the
prologue and also possible to simplify stack representation.

Reviewed By: ricklavoie

Differential Revision: D23806591

fbshipit-source-id: a38cc9c144241910de2acaa3f807d4c53989ba0e