Fix stack representation in stublogue section of prologues
Summary:
When entering prologues, rvmfp() points to the caller's frame and rvmsp() to
the place where callee's frame will eventually be spilled to. Before
DefFrameRelSP, prologues operate in a stublogue context with a fully functional
stack. But since we don't know the delta between rvmfp() and other stack items
statically, we use rvmsp() as the stack base (resumables do a similar thing, as
rvmfp() lives on the heap) and only items below rvmsp() are accessible.

We were not setting this configuration properly, so let's fix it:

1. In genFuncPrologue(), pass the number of TVs alive on the stack below
rvmsp() to the TransContext. This was an oversight from an older diff, as the
correct value was already passed to the prologue_context().

2. In IRGS's constructor, pass the proper offset between rvmsp() and stack base
to the DefRegSP, which is zero. Note that this is different from resumables, as
they get resumed with rvmsp() pointing to the top of the stack, while rvmsp()
in prologues points to the stack base. Also, do not rely on an arbitrary next
opcode to properly setup exception stack boundary.

Reviewed By: ricklavoie

Differential Revision: D23422186

fbshipit-source-id: ef8efad1ffe75260bcd7a63565e552dd8b98971f