Avoid parsing nested expression trees
Summary:
We had an issue parsing constants correctly in Expression Trees, where it would parse the constant as a type specifier. For instance, the following would parse correctly:

```
Code`MY_CONST + 1`;
```

But the following would not

```
Code`MY_CONST`;
Code`1 + MY_CONST`;
```

This is because when we parse `MY_CONST` in the latter examples, we peek at the next token. If the token was a backtick, we would attempt to parse another Expression Tree rather than recognize the backtick as the closing backtick for the existing open Expression Tree.

In order to address this, add a flag to the parser to indicate whether we are in an Expression Tree parsing context. This change means that we will no longer parse nested Expression Trees. So the following would not parse correctly:

```
Code`Code`4``;
```

Reviewed By: Wilfred

Differential Revision: D23721857

fbshipit-source-id: b0f02e51fd98f21d528133a3cb330a22155ff3a4