Error on nested splicing
Summary:
Nested splicing in Expression Trees doesn't make sense.

We already throw errors for splicing not in an Expression Tree context.

```
// Already an error
function test(): void {
  __splice__(4);
}
```
```
// Correct and currently not an error
function test(): void {
  Code`__splice__(4)`;
}
```

We just need to make sure we don't allow that environment to propagate in a splicing context.

```
// Incorrect and currently not an error
function test(): void {
  Code`__splice__(__splice__(4))`;
}
```

This diff makes this later case an error.

This implementation is temporary while we decide on a new syntax. But adding these test cases in so that we remember in the future to redo these tests.

Reviewed By: Wilfred

Differential Revision: D23792648

fbshipit-source-id: e300ba18f5005d37a782bd7a359fa7a26b499d75